[masterbots.ai]refactor:moving chat componets to folder
