+++ apps/masterbots.ai/app/share/[id]/page.tsx
@@ -2,7 +2,7 @@ import { type Metadata } from 'next'
 import { notFound } from 'next/navigation'
 
 import { formatDate } from '@/lib/utils'
-import { ChatList } from '@/components/chat-list'
+import { ChatList } from '@/components/chat/chat-list'
 import { FooterText } from '@/components/footer'
 
 interface SharePageProps {
