+++ apps/masterbots.ai/components/browse-list-item.tsx
@@ -7,7 +7,7 @@ import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import React, { useEffect } from 'react'
 import { BrowseChatMessageList } from './browse-chat-message-list'
-import { ChatAccordion } from './chat-accordion'
+import { ChatAccordion } from '@/components/chat/chat-accordion'
 import { ShortMessage } from './short-message'
 import { IconOpenAI, IconUser } from './ui/icons'
 import { useBrowse } from '@/lib/hooks/use-browse'
