+++ apps/masterbots.ai/components/chat/chat-clickable-text.tsx
