+++ apps/masterbots.ai/components/chat/new-chat.tsx
