+++ apps/masterbots.ai/components/sidebar-actions.tsx
@@ -17,7 +17,7 @@ import {
 } from '@/components/ui/alert-dialog'
 import { Button } from '@/components/ui/button'
 import { IconShare, IconSpinner, IconTrash } from '@/components/ui/icons'
-import { ChatShareDialog } from '@/components/chat-share-dialog'
+import { ChatShareDialog } from '@/components/chat/chat-share-dialog'
 import {
   Tooltip,
   TooltipContent,
