+++ apps/masterbots.ai/components/thread-list.tsx
@@ -4,8 +4,8 @@ import Image from 'next/image'
 import { Thread } from 'mb-genql'
 import React from 'react'
 import { ShortMessage } from './short-message'
-import { ChatAccordion } from './chat-accordion'
-import { ChatList } from './chat-list'
+import { ChatAccordion } from '@/components/chat/chat-accordion'
+import { ChatList } from '@/components/chat/chat-list'
 import { useThread } from '@/lib/hooks/use-thread'
 import { useSidebar } from '@/lib/hooks/use-sidebar'
 import { cn, sleep } from '@/lib/utils'
